Move stuff into a helper file
